+++ lib/augury/exception.rb
@@ -3,7 +3,7 @@
 module Augury
   class TwitterConfigError < StandardError
     def message
-      "No twitter credential configuration found in the augury config"
+      'No twitter credential configuration found in the augury config'
     end
   end
 end