+++ spec/lib/augury/fortune_spec.rb
@@ -1,6 +1,50 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
+require 'open3'
 
 describe Augury::Fortune do
+  context 'fortune' do
+    after do
+      # TODO: better handling of tmp files
+      File.delete('/tmp/boredelonmusk', '/tmp/boredelonmusk.dat') if File.exist? '/tmp/boredelonmusk'
+      File.delete('/tmp/drunkhulk', '/tmp/drunkhulk.dat') if File.exist? '/tmp/drunkhulk'
+    end
+
+    let(:twitter_auth) do
+      {
+        twitter: {
+          'consumer_key' => ENV['TWITTER_CONSUMER_KEY'],
+          'consumer_secret' => ENV['TWITTER_CONSUMER_SECRET'],
+          'access_token' => ENV['TWITTER_ACCESS_TOKEN'],
+          'token_secret' => ENV['TWITTER_ACCESS_TOKEN_SECRET'],
+        },
+      }
+    end
+
+    it 'writes to filesystem' do
+      config = twitter_auth.merge({ count: 3 })
+      augury = Augury::Fortune.new('boredelonmusk', '/tmp/boredelonmusk', config)
+      augury.twitter_setup
+      VCR.use_cassette('twitter_count_3') do
+        augury.retrieve_tweets
+      end
+      augury.write_fortune
+      expect(File).to exist('/tmp/boredelonmusk')
+      expect(File).to exist('/tmp/boredelonmusk.dat')
+    end
+
+    it 'outputs' do
+      config = twitter_auth.merge({ count: 1 })
+      augury = Augury::Fortune.new('drunkhulk', '/tmp/drunkhulk', config)
+      augury.twitter_setup
+      VCR.use_cassette('twitter_count_1') do
+        augury.retrieve_tweets
+      end
+      augury.write_fortune
+      output, res = Open3.capture2('fortune', '/tmp/drunkhulk')
+      expect(res.success?)
+      expect(output).to eq "FUCK THIS!\n"
+    end
+  end
 end